Fix title recursion for frame breaks
If allowframebreaks is used. The Title is repeated n times, with n the
number of used framebreaks. When allowframebreaks is used a short title for
the frame is automatically added. Since this theme renews the frametitle
command and appends the short title to the frametitle, the frametitle is
duplicated for each framebreak resulting in a slide with 4 titles on the
third framebreak.

This patch simply checks for the breakcount and if this is a frame with
multiple breaks the original frametitle command function is used.